code pour la video fini

+++ main.py
@@ -30,7 +30,7 @@ def training(model, X, Y, epoch):
 
         loss = criterion(yPredict, y)
         losses.append(loss)
-
+        print(loss)
         optimizer.zero_grad()
         loss.backward()
         optimizer.step()
@@ -44,39 +44,43 @@ def main():
     # X : Image en niveau de gris
     # Y : composantes a et b de l'image LAB
     # (originale, Xtest, Ytest) = ouvrirImage("test/street-test.jpg", affichage=False)
-    (originale, Xtest, Ytest) = ouvrirImage("test/flower_test.jpg", affichage=False)
-
+    (originale, Xtest, Ytest) = ouvrirImage("test/testimagepourval.jpg", affichage=False)
+    # afficherPrediction(originale, Xtest, Ytest)
     # Telechargement Dataset
     # if not os.path.exists("dataset"):
     #     telechargerDataSet(keywords="landscape", taille=100)
-    (X, Y) = chargerDataset()
 
-    # Creation du CNN
     model = ColorizationNetworkv2()
 
-    # Recuperer un modele existant si besoin
-    # model.load_state_dict(torch.load("models/model_benjamin.pth"))
+    if os.path.isfile('models/model-epoch-4-losses-0.003.pth'):
+        model.load_state_dict(torch.load('models/model-epoch-4-losses-0.003.pth'))
+    else:
 
-    #Deployer le modele sur le gpu
-    model.cuda()
+        (X, Y) = chargerDataset()
+        # Creation du CNN
 
-    # Sinon entrainer un nouveau modele 
-    for epoch in range(EPOCH):
-        training(model, X, Y, epoch)
-    print("entrainement fini")
+        # Recuperer un modele existant si besoin
+        # model.load_state_dict(torch.load("models/model_benjamin.pth"))
 
-    #Sauvegarde du modèle
-    torch.save(model.state_dict(), 'models/model_benjamin.pth')
-    print("sauvegarde du modele")
+        #Deployer le modele sur le gpu
+        model.cuda()
+
+        # Sinon entrainer un nouveau modele 
+        for epoch in range(EPOCH):
+            training(model, X, Y, epoch)
+        print("entrainement fini")
+
+        #Sauvegarde du modèle
+        torch.save(model.state_dict(), 'models/model_benjamin.pth')
+        print("sauvegarde du modele")
 
     # Petit test rapide pour rigoler
     Xtest = Xtest.reshape((1, 1, 256, 256))
-    Xtest_gpu =  Variable(torch.from_numpy(Xtest).float(), volatile=True).cuda()
+    Xtest_gpu =  Variable(torch.from_numpy(Xtest).float(), volatile=True)
     output = model(Xtest_gpu).cpu().detach().numpy()
     
     print("output genere")
     output = output.reshape((256,256,2))
-    output = output * 128
     print("output traite")
     Xtest = Xtest.reshape((256, 256, 1))
     
@@ -84,7 +88,22 @@ def main():
     afficherPrediction(originale, Xtest, output)
 
 
+def get_frame(input):
+    w,h = input.shape[0], input.shape[1]
+    lab = color.rgb2lab(1.0/255*input)
+    ab = lab[:, :, 1:]
+    x = lab[:, :, 0]
+    model = ColorizationNetworkv2()
+    model.load_state_dict(torch.load('models/model-epoch-4-losses-0.003.pth'))
+    model.cuda()
 
-
+    x = x.reshape((1, 1, w, h))
+    x_gpu =  Variable(torch.from_numpy(x).float(), volatile=True).cuda()
+    output = model(x_gpu).cpu().detach().numpy().reshape((w, h, 2)) # output sous la forme ab
+    x = x.reshape(( w,h, 1 ))
+    new_lab = np.concatenate((x, output), axis=2)
+    rgb_output = color.lab2rgb(new_lab)
+    rgb_output *= 255
+    return rgb_output.astype('uint8')
 if __name__ == "__main__":
     main()
